[scide] Move app palette code into separate file
